feat(commands): add joinUp and joinDown command
