+++ docs/api/commands/join-down.md
@@ -1,0 +1,8 @@
+# joinDown
+The `joinDown` command joins the selected block, or if there is a text selection, the closest ancestor block of the selection that can be joined, with the sibling below it. [See also](https://prosemirror.net/docs/ref/#commands.joinDown)
+
+## Usage
+```js
+editor.commands.joinDown()
+```
+
