+++ docs/api/commands/join-up.md
@@ -1,0 +1,8 @@
+# joinUp
+The `joinUp` command joins the selected block, or if there is a text selection, the closest ancestor block of the selection that can be joined, with the sibling above it. [See also](https://prosemirror.net/docs/ref/#commands.joinUp)
+
+## Usage
+```js
+editor.commands.joinUp()
+```
+
