+++ packages/core/src/commands/index.ts
@@ -15,8 +15,7 @@ export * from './focus'
 export * from './forEach'
 export * from './insertContent'
 export * from './insertContentAt'
-export * from './joinBackward'
-export * from './joinForward'
+export * from './join'
 export * from './keyboardShortcut'
 export * from './lift'
 export * from './liftEmptyBlock'
